+++ slpm-server/prisma/migrations/20260805080010_add_release_notes/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "ProductVersion" ADD COLUMN     "releaseNotes" TEXT NOT NULL DEFAULT '';